requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0

# PDF generation
reportlab>=4.0.0

# Configuration and data handling
pyyaml>=6.0
openpyxl>=3.1.0

# Date/time handling
python-dateutil>=2.8.0

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0